drivers: net: zynq_gem: fix BD size for RV64 on Zynq
Zynq uses a 32-bit hardware DMA engine, in which
the size of each buffer descriptor (BD) is inherently configured
as 64-bit (lower 32-bit address of one buffer
and a 32-bit status word of this buffer).
However, when RV64 is adopted, the default BD size
is defined as 128-bit (with an additional higher 32-bit address
of one buffer and a reserved 32-bit word) in this driver.
To overcome such mismatch in BD size when prototyping RV64 on Zynq,
the driver software must ensure that only 64-bit BD format
is leveraged.